+++ R/plot_hist.R
@@ -40,7 +40,8 @@
 #'
 #'@export
 
-  plot_hist <- function(object, xCol = 1, wCol = "all", base = FALSE, x_limits,                            displ = TRUE, binwidth, displLines = FALSE){
+  plot_hist <- function(object, xCol = 1, wCol = "all", base = FALSE, x_limits,                            
+                        displ = TRUE, binwidth, displLines = FALSE){
    if (!is.SWIM(object)) stop("Object not of class SWIM")
    if (anyNA(object$x)) warning("x contains NA")
    x_data <- get_data(object)[ , xCol]