+++ NEWS.md
@@ -1,5 +1,32 @@
 
-# SWIM 0.2.1
+# SWIM 0.2.1.9000 - develop version
+
+## Major changes: Additional functions and features
+
+ - `plot_quantile()`:
+    * the function plots the empirical quantile of model components, subject to 
+      scenario weights.
+
+ - `plot_weights()`:
+    * the function plots the scenario weights of a stressed model agains 
+      model components.
+
+ - `stress_moment()`:
+    * add parameter "normalise" that allows to linearly normalise the values
+    called by `nleqslv`.
+    * the function prints a table with the required and achieved moments and the absolute and relative error.
+
+ - `stress_VaR_ES()`:
+    * add parameter "normalise" that allows to linearly normalise the values 
+    before `uniroot` is applied.
+
+
+## Minor changes
+
+ - fix bug in merging different stress objects.
+
+
+# SWIM 0.2.1 - current version on CRAN
 
 ## Minor changes